policies-gates

+++ app/Http/Controllers/ServiceController.php
@@ -14,6 +14,7 @@ class ServiceController extends Controller
     }
     public function create()
     {
+        $this->authorize('create', Service::class);
         $services = Service::all();
         return view("/back/services/create", compact("services"));
     }
@@ -44,6 +45,7 @@ class ServiceController extends Controller
     public function update($id, Request $request)
     {
         $service = Service::find($id);
+        $this->authorize('update', $service);
         $request->validate([
             'icone'=> 'required',
             'title'=> 'required',
@@ -58,6 +60,7 @@ class ServiceController extends Controller
     public function destroy($id)
     {
         $service = Service::find($id);
+        $this->authorize('detete', $service);
         $service->delete();
         return redirect()->back()->with("message", "Service successfuly deleted !");
     }
